Fix Keycloak OpenShift scenarios on IBM Power Z
