+++ security/keycloak-oidc-client-reactive-extended/src/test/java/io/quarkus/ts/security/keycloak/oidcclient/reactive/extended/SecurityClaimsIT.java
@@ -1,19 +1,34 @@
 package io.quarkus.ts.security.keycloak.oidcclient.reactive.extended;
 
+import static io.quarkus.test.bootstrap.KeycloakService.DEFAULT_REALM;
+import static io.quarkus.test.bootstrap.KeycloakService.DEFAULT_REALM_BASE_PATH;
+import static io.quarkus.test.bootstrap.KeycloakService.DEFAULT_REALM_FILE;
+import static io.quarkus.ts.security.keycloak.oidcclient.reactive.extended.OidcItUtils.createToken;
 import static io.restassured.RestAssured.given;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.Test;
 
+import io.quarkus.test.bootstrap.KeycloakService;
+import io.quarkus.test.bootstrap.RestService;
 import io.quarkus.test.scenarios.QuarkusScenario;
+import io.quarkus.test.services.KeycloakContainer;
+import io.quarkus.test.services.QuarkusApplication;
 
 @QuarkusScenario
-public class SecurityClaimsIT extends BaseOidcIT {
+public class SecurityClaimsIT {
 
     private static final String SECURED_PATH = "/secured";
     private static final String CLAIMS_FROM_BEANS_PATH = "/getClaimsFromBeans";
     private static final String CLAIMS_FROM_TOKEN_PATH = "/getClaimsFromToken";
+
+    @KeycloakContainer(command = { "start-dev", "--import-realm", "--features=token-exchange" })
+    static KeycloakService keycloak = new KeycloakService(DEFAULT_REALM_FILE, DEFAULT_REALM, DEFAULT_REALM_BASE_PATH);
+
+    @QuarkusApplication
+    static RestService app = new RestService()
+            .withProperty("quarkus.oidc.auth-server-url", () -> keycloak.getRealmUrl());
 
     @Test
     public void verifySecuredEndpointIsProtected() {
@@ -39,7 +54,7 @@ public class SecurityClaimsIT extends BaseOidcIT {
 
     private String getClaimsInstancesFromPath(String path) {
         return given()
-                .auth().preemptive().oauth2(createToken())
+                .auth().preemptive().oauth2(createToken(keycloak))
                 .get(SECURED_PATH + path)
                 .then().statusCode(HttpStatus.SC_OK)
                 .extract().asString();
