+++ security/keycloak/src/test/java/io/quarkus/ts/security/keycloak/OpenShiftMultiMechanismAuthenticationIT.java
@@ -1,10 +1,20 @@
 package io.quarkus.ts.security.keycloak;
 
+import static io.quarkus.test.bootstrap.KeycloakService.DEFAULT_REALM;
+import static io.quarkus.test.bootstrap.KeycloakService.DEFAULT_REALM_BASE_PATH;
+import static io.quarkus.test.bootstrap.KeycloakService.DEFAULT_REALM_FILE;
+
 import org.junit.jupiter.api.Tag;
 
+import io.quarkus.test.bootstrap.KeycloakService;
 import io.quarkus.test.scenarios.OpenShiftScenario;
+import io.quarkus.test.services.KeycloakContainer;
 
 @OpenShiftScenario
 @Tag("QUARKUS-5617")
-public class OpenShiftMultiMechanismAuthenticationIT extends MultiMechanismAuthenticationIT {
+public class OpenShiftMultiMechanismAuthenticationIT extends AbstractMultiMechanismAuthenticationIT {
+
+    @KeycloakContainer(command = { "start-dev", "--import-realm" }, image = "${rhbk.image}")
+    static KeycloakService keycloak = new KeycloakService(DEFAULT_REALM_FILE, DEFAULT_REALM, DEFAULT_REALM_BASE_PATH);
+
 }
